requirements.txt
djangorestframework==3.12.4
Pillow==8.4.0
requests==2.26.0